pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gasv"
version = "0.1.0"
description = "GitHub Actions Security Validator - static analysis for CI/CD workflow files"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
gasv = "gasv.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
